refactor: move to new api route usage
